cleaned up Bytetypes
how are we going to actually handle writing these to `Bytes.t`? for
static types we can ensure there is enough room but for streams we can't
be sure and there is no way to go "back in time" to try writing the
elements to a bigger `Bytes.t`. perhaps we should create an
automatically resizing type that wraps `Bytes.t`? there is no way we are
fast to begin with so it couldn't hurt.